refactor(setup): split wizard.go into model, update, view, and commands
