--- cmd/setup/wizard.go
@@ -1,792 +1,0 @@
-package setup
-
-import (
-	"fmt"
-	"os"
-	"strings"
-
-	"github.com/charmbracelet/bubbles/list"
-	"github.com/charmbracelet/bubbles/spinner"
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-
-	"github.com/launchdarkly/ldcli/cmd/cliflags"
-	"github.com/launchdarkly/ldcli/internal/analytics"
-	"github.com/launchdarkly/ldcli/internal/errors"
-	"github.com/launchdarkly/ldcli/internal/setup"
-)
-
-type wizardStep int
-
-const (
-	stepSelectProject wizardStep = iota
-	stepSelectEnvironment
-	stepDetect
-	stepSelectSDK
-	stepPlan
-	stepInstall
-	stepCreateFlag
-	stepInit
-	stepWaitForApp
-	stepVerify
-	stepDone
-)
-
-type wizardModel struct {
-	analyticsTrackerFn analytics.TrackerFn
-	svc                setup.Service
-	auth               setup.Auth
-
-	step    wizardStep
-	spinner spinner.Model
-	err     error
-	width   int
-	height  int
-
-	// data gathered through the flow
-	projects     []projectItem
-	environments []envItem
-	projectList  list.Model
-	envList      list.Model
-	sdkList      list.Model
-
-	selectedProject string
-	selectedEnv     string
-	sdkKey          string
-	clientSideID    string
-	mobileKey       string
-
-	detectComplete     bool   // detection (run once at launch) has finished
-	detectedSDKID      string // detected SDK id, cached from the one-time detection ("" if none)
-	detectedEntryPoint string
-	detectResult       *setup.DetectResult
-	detectedSDK        *sdkItem // the auto-detected SDK, shown in its own panel; nil if detection failed
-	sdkFocus           int      // on the SDK screen: 0 = detected panel, 1 = the list of other SDKs
-	planInstallCmd     string   // install command previewed on the plan screen
-	planAlready        bool     // whether the SDK is already installed (previewed on the plan screen)
-	installResult      *setup.InstallResult
-	flagKey            string
-	initResult         *setup.InitResult
-	verifyResult       *setup.VerifyResult
-
-	quitting bool
-}
-
-type sdkItem struct {
-	id       string
-	language string
-	name     string
-}
-
-func (s sdkItem) Title() string {
-	if setup.RequiresManualInstall(s.id) {
-		return s.name + " (manual install)"
-	}
-	return s.name
-}
-func (s sdkItem) Description() string { return s.language }
-func (s sdkItem) FilterValue() string { return s.name }
-
-type projectItem struct {
-	key  string
-	name string
-}
-
-func (p projectItem) Title() string       { return p.name }
-func (p projectItem) Description() string { return p.key }
-func (p projectItem) FilterValue() string { return p.name }
-
-type envItem struct {
-	key  string
-	name string
-}
-
-func (e envItem) Title() string       { return e.name }
-func (e envItem) Description() string { return e.key }
-func (e envItem) FilterValue() string { return e.name }
-
-// messages
-type projectsFetchedMsg struct{ projects []projectItem }
-type envsFetchedMsg struct{ environments []envItem }
-type envDetailsFetchedMsg struct {
-	sdkKey       string
-	clientSideID string
-	mobileKey    string
-}
-type detectDoneMsg struct{ result *setup.DetectResult }
-type detectFailedMsg struct{}
-type installDoneMsg struct{ result *setup.InstallResult }
-type flagCreatedMsg struct{ key string }
-type initDoneMsg struct{ result *setup.InitResult }
-type verifyDoneMsg struct{ result *setup.VerifyResult }
-type wizardErrMsg struct{ err error }
-
-func runSetupWizard(
-	analyticsTrackerFn analytics.TrackerFn,
-	svc setup.Service,
-) func(*cobra.Command, []string) error {
-	return func(cmd *cobra.Command, args []string) error {
-		// Pre-flight: the wizard's first action is an authenticated API call, so
-		// bail early with clear guidance rather than dumping a raw 401 mid-TUI.
-		if viper.GetString(cliflags.AccessTokenFlag) == "" {
-			return errors.NewError("It looks like you're not logged in yet.\n\nRun `ldcli login` to authenticate, then run `ldcli setup` again.\n(Or pass --access-token, or set LD_ACCESS_TOKEN.)")
-		}
-
-		s := spinner.New()
-		s.Spinner = spinner.Dot
-
-		m := wizardModel{
-			analyticsTrackerFn: analyticsTrackerFn,
-			svc:                svc,
-			auth: setup.Auth{
-				AccessToken: viper.GetString(cliflags.AccessTokenFlag),
-				BaseURI:     viper.GetString(cliflags.BaseURIFlag),
-			},
-			step:    stepSelectProject,
-			spinner: s,
-		}
-
-		p := tea.NewProgram(m, tea.WithAltScreen())
-		_, err := p.Run()
-		return err
-	}
-}
-
-func (m wizardModel) Init() tea.Cmd {
-	// Detect the project once, up front, so navigating the flow never re-runs it.
-	return tea.Batch(m.spinner.Tick, m.fetchProjects(), m.runDetect())
-}
-
-func (m wizardModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tea.WindowSizeMsg:
-		m.width = msg.Width
-		m.height = msg.Height
-
-	case tea.KeyMsg:
-		switch msg.String() {
-		case "ctrl+c":
-			m.quitting = true
-			return m, tea.Quit
-		case "q", "esc":
-			if m.isFiltering() {
-				break // let the list receive 'q' / clear its filter
-			}
-			m.quitting = true
-			return m, tea.Quit
-		case "left", "h":
-			if m.isFiltering() {
-				break // let the list receive the key as filter input
-			}
-			return m.handleBack()
-		case "enter":
-			return m.handleEnter()
-		}
-
-	case projectsFetchedMsg:
-		m.projects = msg.projects
-		items := make([]list.Item, len(msg.projects))
-		for i, p := range msg.projects {
-			items[i] = p
-		}
-		delegate := list.NewDefaultDelegate()
-		m.projectList = list.New(items, delegate, m.width, m.height-4)
-		m.projectList.Title = "Select a project:"
-		m.projectList.SetShowStatusBar(false)
-		return m, nil
-
-	case envsFetchedMsg:
-		m.environments = msg.environments
-		items := make([]list.Item, len(msg.environments))
-		for i, e := range msg.environments {
-			items[i] = e
-		}
-		delegate := list.NewDefaultDelegate()
-		m.envList = list.New(items, delegate, m.width, m.height-4)
-		m.envList.Title = "Select an environment:"
-		m.envList.SetShowStatusBar(false)
-		return m, nil
-
-	case envDetailsFetchedMsg:
-		m.sdkKey = msg.sdkKey
-		m.clientSideID = msg.clientSideID
-		m.mobileKey = msg.mobileKey
-		// Detection was kicked off at launch; go straight to the SDK screen if
-		// it's already done, otherwise show a brief wait until it lands.
-		if m.detectComplete {
-			m.enterSDKStep()
-		} else {
-			m.step = stepDetect
-		}
-		return m, nil
-
-	case detectFailedMsg:
-		m.detectComplete = true
-		m.detectedSDKID = ""
-		if m.step == stepDetect {
-			m.enterSDKStep()
-		}
-		return m, nil
-
-	case detectDoneMsg:
-		m.detectComplete = true
-		m.detectedSDKID = msg.result.SDKID
-		m.detectedEntryPoint = msg.result.EntryPoint
-		if m.step == stepDetect {
-			m.enterSDKStep()
-		}
-		return m, nil
-
-	case installDoneMsg:
-		m.installResult = msg.result
-		m.step = stepCreateFlag
-		return m, m.runCreateFlag()
-
-	case flagCreatedMsg:
-		m.flagKey = msg.key
-		m.step = stepInit
-		return m, m.runInit()
-
-	case initDoneMsg:
-		m.initResult = msg.result
-		// Skip the live verify if init didn't inject runnable code, or if the SDK
-		// wasn't actually installed (auto-install failed) — the app can't connect.
-		if !msg.result.Success || (m.installResult != nil && m.installResult.Failed) {
-			m.step = stepDone
-			return m, nil
-		}
-		m.step = stepWaitForApp
-		return m, nil
-
-	case verifyDoneMsg:
-		m.verifyResult = msg.result
-		m.step = stepDone
-		return m, nil
-
-	case wizardErrMsg:
-		m.err = msg.err
-		return m, nil
-
-	case spinner.TickMsg:
-		var cmd tea.Cmd
-		m.spinner, cmd = m.spinner.Update(msg)
-		return m, cmd
-	}
-
-	// delegate to list models
-	var cmd tea.Cmd
-	switch m.step {
-	case stepSelectProject:
-		if len(m.projects) > 0 {
-			m.projectList, cmd = m.projectList.Update(msg)
-		}
-	case stepSelectEnvironment:
-		if len(m.environments) > 0 {
-			m.envList, cmd = m.envList.Update(msg)
-		}
-	case stepSelectSDK:
-		// Two panels when a detected SDK is shown: the detected panel (focus 0)
-		// and the list of other SDKs (focus 1). Arrows move focus between them.
-		if m.detectedSDK != nil {
-			if km, ok := msg.(tea.KeyMsg); ok {
-				switch km.String() {
-				case "down", "tab", "j":
-					if m.sdkFocus == 0 {
-						m.sdkFocus = 1
-						m.sdkList.SetDelegate(sdkDelegate(true))
-						return m, nil
-					}
-				case "up", "shift+tab", "k":
-					if m.sdkFocus == 1 && m.sdkList.Index() == 0 {
-						m.sdkFocus = 0
-						m.sdkList.SetDelegate(sdkDelegate(false))
-						return m, nil
-					}
-				}
-			}
-			if m.sdkFocus == 1 && m.sdkList.Items() != nil {
-				m.sdkList, cmd = m.sdkList.Update(msg)
-			}
-		} else if m.sdkList.Items() != nil {
-			m.sdkList, cmd = m.sdkList.Update(msg)
-		}
-	}
-	return m, cmd
-}
-
-// isFiltering reports whether the current step's list is in filter-typing mode,
-// so keys like esc/q are left for the list instead of triggering back/quit.
-func (m wizardModel) isFiltering() bool {
-	switch m.step {
-	case stepSelectProject:
-		return m.projectList.FilterState() == list.Filtering
-	case stepSelectEnvironment:
-		return m.envList.FilterState() == list.Filtering
-	case stepSelectSDK:
-		return m.sdkList.FilterState() == list.Filtering
-	}
-	return false
-}
-
-// enterSDKStep builds the SDK-selection screen from the cached one-time
-// detection result and switches to it. Rebuilding the list is cheap and uses
-// the current width; detection itself is never re-run.
-func (m *wizardModel) enterSDKStep() {
-	if id := m.detectedSDKID; id != "" {
-		if det, ok := findKnownSDK(id); ok {
-			m.detectedSDK = &det
-			m.sdkFocus = 0
-			m.sdkList = m.newSDKList(sdkItemsExcept(det.id), "Other SDKs:", false)
-			m.step = stepSelectSDK
-			return
-		}
-	}
-	m.detectedSDK = nil
-	m.sdkFocus = 1
-	m.sdkList = m.newSDKList(sdkItemsExcept(""), "Select your SDK:", true)
-	m.step = stepSelectSDK
-}
-
-// handleBack returns to the previous selection so the user can change the
-// project, environment, or SDK.
-func (m wizardModel) handleBack() (tea.Model, tea.Cmd) {
-	switch m.step {
-	case stepSelectEnvironment:
-		m.step = stepSelectProject
-	case stepSelectSDK:
-		m.step = stepSelectEnvironment
-	case stepPlan:
-		m.step = stepSelectSDK
-	}
-	return m, nil
-}
-
-func (m wizardModel) handleEnter() (tea.Model, tea.Cmd) {
-	switch m.step {
-	case stepSelectProject:
-		if len(m.projects) == 0 {
-			return m, nil
-		}
-		selected, ok := m.projectList.SelectedItem().(projectItem)
-		if !ok {
-			return m, nil
-		}
-		m.selectedProject = selected.key
-		m.step = stepSelectEnvironment
-		return m, m.fetchEnvironments()
-
-	case stepSelectEnvironment:
-		if len(m.environments) == 0 {
-			return m, nil
-		}
-		selected, ok := m.envList.SelectedItem().(envItem)
-		if !ok {
-			return m, nil
-		}
-		m.selectedEnv = selected.key
-		return m, m.fetchEnvDetails()
-
-	case stepSelectSDK:
-		var chosen sdkItem
-		if m.detectedSDK != nil && m.sdkFocus == 0 {
-			chosen = *m.detectedSDK
-		} else {
-			selected, ok := m.sdkList.SelectedItem().(sdkItem)
-			if !ok {
-				return m, nil
-			}
-			chosen = selected
-		}
-		m.detectResult = &setup.DetectResult{
-			SDKID:      chosen.id,
-			Language:   chosen.language,
-			EntryPoint: m.detectedEntryPoint,
-		}
-		// Compute the plan preview shown before any action is taken.
-		args, _ := setup.InstallArgs(chosen.id, "")
-		m.planInstallCmd = strings.Join(args, " ")
-		if dir, err := os.Getwd(); err == nil {
-			m.planAlready = setup.IsInstalled(dir, chosen.id)
-		}
-		m.step = stepPlan
-		return m, nil
-
-	case stepPlan:
-		m.step = stepInstall
-		return m, m.runInstall()
-
-	case stepWaitForApp:
-		m.step = stepVerify
-		return m, m.runVerify()
-	}
-	return m, nil
-}
-
-// quitHint is appended to terminal (done) screens so the user knows how to exit.
-var quitHint = "\n" + mutedStyle.Render("Press q to quit.") + "\n"
-
-func (m wizardModel) View() string {
-	if m.quitting {
-		return ""
-	}
-
-	if m.err != nil {
-		return titleStyle.Render("Error") + "\n\n" + m.err.Error() + "\n\nPress ctrl+c to quit."
-	}
-
-	switch m.step {
-	case stepSelectProject:
-		if len(m.projects) == 0 {
-			return m.spinner.View() + " Loading projects..."
-		}
-		return m.projectList.View() + "\n" + mutedStyle.Render("esc quit")
-
-	case stepSelectEnvironment:
-		if len(m.environments) == 0 {
-			return m.spinner.View() + " Loading environments..."
-		}
-		return m.envList.View() + "\n" + mutedStyle.Render("← back · esc quit")
-
-	case stepDetect:
-		return m.spinner.View() + " Detecting project type..."
-
-	case stepSelectSDK:
-		return m.sdkSelectView()
-
-	case stepPlan:
-		return m.planView()
-
-	case stepInstall:
-		return m.spinner.View() + " Installing SDK..."
-
-	case stepCreateFlag:
-		return m.spinner.View() + " Creating feature flag..."
-
-	case stepInit:
-		return m.spinner.View() + " Injecting initialization code..."
-
-	case stepWaitForApp:
-		return titleStyle.Render("Start your application") + "\n\n" +
-			"SDK initialization code has been injected into:\n" +
-			"  " + m.initResult.FilePath + "\n\n" +
-			"Please start your application now, then press Enter to verify the connection.\n"
-
-	case stepVerify:
-		return m.spinner.View() + " Waiting for SDK to connect..."
-
-	case stepDone:
-		if m.installResult != nil && m.installResult.Failed {
-			body := titleStyle.Render("Manual install needed") + "\n\n" +
-				m.wrap("The SDK couldn't be installed automatically. Install it yourself with:") + "\n\n" +
-				code(m.installResult.Command) + "\n\n"
-			if m.installResult.FailureReason != "" {
-				body += m.wrap("Reason: "+m.installResult.FailureReason) + "\n\n"
-			}
-			if m.initResult != nil && m.initResult.Success {
-				body += m.wrap(fmt.Sprintf("Initialization code was added to %s.", m.initResult.FilePath)) + "\n"
-			} else if m.initResult != nil && m.initResult.Snippet != "" {
-				body += m.wrap(fmt.Sprintf("Then add this initialization code to %s:", m.initResult.FilePath)) +
-					"\n\n" + code(m.initResult.Snippet) + "\n"
-			}
-			body += "\n" + m.wrap(fmt.Sprintf("Flag %q was created in project %q.", m.flagKey, m.selectedProject)) + "\n"
-			return body + quitHint
-		}
-		if m.initResult != nil && !m.initResult.Success {
-			body := titleStyle.Render("Manual SDK setup required") + "\n\n"
-			if m.initResult.Snippet != "" {
-				body += m.wrap(fmt.Sprintf("Add the following %s initialization code to %s:", m.initResult.SDKID, m.initResult.FilePath)) +
-					"\n\n" + code(m.initResult.Snippet) + "\n\n"
-			} else {
-				body += fmt.Sprintf("No initialization template is available for %s.\n", m.initResult.SDKID)
-			}
-			return body +
-				fmt.Sprintf("Follow the setup guide at: %s\n\n", m.initResult.DocsURL) +
-				fmt.Sprintf("Flag %q has been created in project %q.\n", m.flagKey, m.selectedProject) +
-				"Once you've initialized the SDK manually, your flag will be ready to use.\n" +
-				quitHint
-		}
-		if m.verifyResult != nil && m.verifyResult.Active && m.detectResult != nil {
-			appHost := strings.TrimRight(viper.GetString(cliflags.BaseURIFlag), "/")
-			return titleStyle.Render("Setup complete!") + "\n\n" +
-				fmt.Sprintf("Your %s SDK is connected to LaunchDarkly.\n", m.detectResult.SDKID) +
-				fmt.Sprintf("Flag %q is ready to use.\n\n", m.flagKey) +
-				fmt.Sprintf("You can now toggle your flag at %s/projects/%s/flags/%s/targeting?env=%s\n", appHost, m.selectedProject, m.flagKey, m.selectedEnv) +
-				quitHint
-		}
-		return titleStyle.Render("Verification timed out") + "\n\n" +
-			"The SDK did not report as active within the timeout period.\n" +
-			"Make sure your application is running and try again.\n" +
-			quitHint
-	}
-
-	return ""
-}
-
-// findKnownSDK returns the sdkItem for the given SDK id, if it is one we know.
-func findKnownSDK(id string) (sdkItem, bool) {
-	for _, sdk := range setup.KnownSDKs {
-		if sdk.ID == id {
-			return sdkItem{id: sdk.ID, language: sdk.Language, name: sdk.Name}, true
-		}
-	}
-	return sdkItem{}, false
-}
-
-// sdkItemsExcept returns all known SDKs as list items, omitting the given id.
-func sdkItemsExcept(exclude string) []list.Item {
-	items := make([]list.Item, 0, len(setup.KnownSDKs))
-	for _, sdk := range setup.KnownSDKs {
-		if sdk.ID == exclude {
-			continue
-		}
-		items = append(items, sdkItem{id: sdk.ID, language: sdk.Language, name: sdk.Name})
-	}
-	return items
-}
-
-// sdkBoxWidth is the shared width for the detected panel and the SDK list box,
-// so both areas line up.
-func (m wizardModel) sdkBoxWidth() int {
-	w := m.width - 4
-	if w > 72 {
-		w = 72
-	}
-	if w < 20 { // never wider than a very narrow terminal can show
-		w = 20
-	}
-	return w
-}
-
-// wrap reflows prose to the terminal width so it doesn't overflow narrow
-// terminals. Code snippets are rendered raw (not passed through here).
-func (m wizardModel) wrap(s string) string {
-	return wrapText(s, m.width)
-}
-
-// sdkDelegate returns the list row renderer. When the list isn't the focused
-// area, the selected row is styled like a normal row so it doesn't look active
-// while the detected-SDK panel holds focus.
-func sdkDelegate(focused bool) list.DefaultDelegate {
-	d := list.NewDefaultDelegate()
-	if !focused {
-		d.Styles.SelectedTitle = d.Styles.NormalTitle
-		d.Styles.SelectedDesc = d.Styles.NormalDesc
-	}
-	return d
-}
-
-// newSDKList builds the list model for the SDK selection screen.
-func (m wizardModel) newSDKList(items []list.Item, title string, focused bool) list.Model {
-	h := m.height - 12
-	if h < 3 {
-		h = 3
-	}
-	l := list.New(items, sdkDelegate(focused), m.sdkBoxWidth()-2, h)
-	l.Title = title
-	l.Styles.Title = headerStyle // match the detected-SDK panel header, not the default title bar
-	l.SetShowStatusBar(false)
-	l.SetShowHelp(false) // we render a single key hint inside the box instead
-	return l
-}
-
-// sdkSelectView renders the SDK selection screen. When an SDK was auto-detected
-// it shows two areas: an "identified" panel on top and the list of other SDKs
-// below; the focused area is highlighted. When detection failed, only the list
-// is shown.
-func (m wizardModel) sdkSelectView() string {
-	hint := mutedStyle.Render("↑/↓ move · enter select · ← back · esc quit")
-	catalog := mutedStyle.Render("Don't see your language? All LaunchDarkly SDKs: https://launchdarkly.com/docs/sdk")
-
-	if m.detectedSDK == nil {
-		listBox := box(true, m.sdkBoxWidth()).Render(m.sdkList.View() + "\n" + hint)
-		return listBox + "\n" + catalog
-	}
-
-	boxW := m.sdkBoxWidth()
-	panelStyle := box(m.sdkFocus == 0, boxW)
-	listStyle := box(m.sdkFocus == 1, boxW)
-
-	// Point to the detected SDK when its panel is focused, matching the list's cursor.
-	label := fmt.Sprintf("%s  (%s)", m.detectedSDK.name, m.detectedSDK.language)
-	if setup.RequiresManualInstall(m.detectedSDK.id) {
-		label += " — manual install"
-	}
-	pointer := "  "
-	if m.sdkFocus == 0 {
-		pointer, label = selectedStyle.Render("❯ "), selectedStyle.Render(label)
-	}
-	panel := panelStyle.Render(
-		headerStyle.Render("We identified this as your SDK") + "\n" +
-			pointer + label + "\n" +
-			mutedStyle.Render("Press Enter to use it"))
-
-	listBox := listStyle.Render(m.sdkList.View() + "\n" + hint)
-
-	return panel + "\n\n" + listBox + "\n" + catalog
-}
-
-// planView lists the steps setup will take, before any of them run, so the user
-// knows what's about to happen and can confirm.
-func (m wizardModel) planView() string {
-	if m.detectResult == nil {
-		return ""
-	}
-	name := m.detectResult.SDKID
-	if nm, ok := findKnownSDK(m.detectResult.SDKID); ok {
-		name = nm.name
-	}
-
-	var steps []string
-	add := func(s string) {
-		steps = append(steps, selectedStyle.Render(fmt.Sprintf("%d.", len(steps)+1))+" "+s)
-	}
-
-	switch {
-	case m.planAlready:
-		add(fmt.Sprintf("Install the %s SDK — %s", name, mutedStyle.Render("already installed, will skip")))
-	case m.planInstallCmd != "":
-		add(fmt.Sprintf("Install the %s SDK — %s", name, mutedStyle.Render(m.planInstallCmd)))
-	default:
-		add(fmt.Sprintf("Add the %s SDK %s", name, mutedStyle.Render("(manual install)")))
-	}
-	add(fmt.Sprintf("Create a feature flag in %s / %s", m.selectedProject, m.selectedEnv))
-	if setup.InjectsInPlace(m.detectResult.SDKID) {
-		// Say when the entry file does not exist yet: a file we create is not loaded
-		// by the project, so the user needs the chance to back out and point us at
-		// the real entry point.
-		if m.detectResult.EntryPointExists {
-			add(fmt.Sprintf("Add initialization code to %s", m.detectResult.EntryPoint))
-		} else {
-			add(fmt.Sprintf("Create %s with initialization code %s",
-				m.detectResult.EntryPoint,
-				mutedStyle.Render("(no entry file found — check this is where your app starts)")))
-		}
-		add("Verify the SDK connects to LaunchDarkly")
-	} else {
-		add("Show initialization code for you to add")
-	}
-
-	return headerStyle.Render("Here's what setup will do:") + "\n\n" +
-		strings.Join(steps, "\n") + "\n\n" +
-		mutedStyle.Render("Enter continue · ← back · esc quit")
-}
-
-// Commands that perform async work. Each is a thin tea.Cmd adapter over the
-// orchestration service: it calls a step method and maps the result or error
-// onto a wizard message. All API/filesystem work and business rules live in
-// internal/setup.Service.
-
-func (m wizardModel) fetchProjects() tea.Cmd {
-	return func() tea.Msg {
-		ps, err := m.svc.ListProjects(m.auth)
-		if err != nil {
-			return wizardErrMsg{err: err}
-		}
-		projects := make([]projectItem, len(ps))
-		for i, p := range ps {
-			projects[i] = projectItem{key: p.Key, name: p.Name}
-		}
-		return projectsFetchedMsg{projects: projects}
-	}
-}
-
-func (m wizardModel) fetchEnvironments() tea.Cmd {
-	return func() tea.Msg {
-		es, err := m.svc.ListEnvironments(m.auth, m.selectedProject)
-		if err != nil {
-			return wizardErrMsg{err: err}
-		}
-		envs := make([]envItem, len(es))
-		for i, e := range es {
-			envs[i] = envItem{key: e.Key, name: e.Name}
-		}
-		return envsFetchedMsg{environments: envs}
-	}
-}
-
-func (m wizardModel) fetchEnvDetails() tea.Cmd {
-	return func() tea.Msg {
-		keys, err := m.svc.EnvKeys(m.auth, m.selectedProject, m.selectedEnv)
-		if err != nil {
-			return wizardErrMsg{err: err}
-		}
-		return envDetailsFetchedMsg{
-			sdkKey:       keys.SDKKey,
-			clientSideID: keys.ClientSideID,
-			mobileKey:    keys.MobileKey,
-		}
-	}
-}
-
-func (m wizardModel) runDetect() tea.Cmd {
-	return func() tea.Msg {
-		dir, err := os.Getwd()
-		if err != nil {
-			return wizardErrMsg{err: err}
-		}
-		result, err := m.svc.Detect(dir)
-		if err != nil {
-			return detectFailedMsg{}
-		}
-		return detectDoneMsg{result: result}
-	}
-}
-
-func (m wizardModel) runInstall() tea.Cmd {
-	return func() tea.Msg {
-		dir, err := os.Getwd()
-		if err != nil {
-			return wizardErrMsg{err: err}
-		}
-		result, err := m.svc.Install(dir, m.detectResult)
-		if err != nil {
-			// Don't dead-end the interactive flow on a failed auto-install (e.g.
-			// Ruby gem perms, no network): surface the command to run by hand.
-			args, _ := setup.InstallArgs(m.detectResult.SDKID, m.detectResult.PackageManager)
-			return installDoneMsg{result: &setup.InstallResult{
-				SDKID:         m.detectResult.SDKID,
-				Command:       strings.Join(args, " "),
-				Failed:        true,
-				FailureReason: err.Error(),
-			}}
-		}
-		return installDoneMsg{result: result}
-	}
-}
-
-func (m wizardModel) runCreateFlag() tea.Cmd {
-	return func() tea.Msg {
-		key, err := m.svc.CreateFlag(m.auth, m.selectedProject, "my-new-flag", "My New Flag")
-		if err != nil {
-			return wizardErrMsg{err: err}
-		}
-		return flagCreatedMsg{key: key}
-	}
-}
-
-func (m wizardModel) runInit() tea.Cmd {
-	return func() tea.Msg {
-		cfg := setup.InitConfig{
-			SDKKey:       m.sdkKey,
-			ClientSideID: m.clientSideID,
-			MobileKey:    m.mobileKey,
-			FlagKey:      m.flagKey,
-		}
-		result, err := m.svc.Inject(m.detectResult.SDKID, m.detectResult.EntryPoint, cfg)
-		if err != nil {
-			return wizardErrMsg{err: err}
-		}
-		return initDoneMsg{result: result}
-	}
-}
-
-func (m wizardModel) runVerify() tea.Cmd {
-	return func() tea.Msg {
-		result, err := m.svc.Verify(m.auth, m.selectedProject, m.selectedEnv)
-		if err != nil {
-			return wizardErrMsg{err: err}
-		}
-		return verifyDoneMsg{result: result}
-	}
-}
